2D Particle System Demo
Description: Demos 4 ops which can be toggled on and off (and otherwise
modified in some cases) with button or mouse presses
Many Status... (Stati?) Are displayed at the bottom of the sketch.

A. Particles are generated based on the content of an image.
B. Translated by Perlin Noise within a rectangle.
C. Accelerate towards a line.